fix(build): add isPitchBlack to MonetConfiguration, fix type mismatches in manual screens

+++ app/src/main/java/dev/aurakai/auraframefx/domains/aura/ui/screens/manual/LaunchMatrixManualScreen.kt
@@ -119,7 +119,7 @@ fun LaunchMatrixManualScreen(
                     onValueChange = {
                         viewModel.updateLauncherConfig(
                             context,
-                            config.copy(iconSize = it / 100f)
+                            config.copy(iconSize = (it / 100f).toInt())
                         )
                     }
                 )
